app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
motor-asyncio
emergentintegrations==0.1.0
